Subscriptions problem solved
I solved the bug in the screen transition that I mentioned yesterday and the problem of the subscriptions not appearing at the top of the screen, but now only the first image appears even though I have entered more than one subscription image. However, I discovered a new bug, when I pull it down on the subscription page, it sends it back to the main screen, I will try to solve it.