I wrote the xml code of the liked videos and integrated it into the main activity.
To make the favorite imageview, I added an icon and text on it and created a new Adapter class. After defining my XML views with the likesVideo class, I used the json type data so that the image I wanted came after my image in the main activity, and I wrote the necessary codes for this in my main activity.