I edited the playlist xml and tried to finalize it, but I think the icons I added on top were small, I will rearrange it. What I'm working on right now, I'm trying to show these two different layouts in a single recyler view. It's not over yet, there are many missing but I will take care of it
I edited the playlist xml and tried to finalize it, but I think the icons I added on top were small, I will rearrange it. What I'm working on right now, I'm trying to show these two different layouts in a single recyler view. It's not over yet, there are many missing but I will take care of it