There are problems because I got the data from a fake database and the project was very disorganized. I still managed to make the floating screen
Hi, I decided to take a break from this project, I've been working on it for weeks and I realized that I have a lot of shortcomings and I realized that I need to make a new project to make up for these shortcomings. Because in this project, I had a lot of problems because I didn't pull the data from the db but I got the data from fake db myself. See you in new projects